refactor(api): split api.py into a modular api/ package
Move the 1934-line AjaxRestApi god-class into a package: api/_base.py holds
AjaxRestClientBase (session, auth login/2FA/refresh/recover, retry/backoff,
_request/_request_no_response, rate-limit, in-memory caches, typed exceptions
and the module constants); the domain endpoint wrappers move verbatim into
_hubs / _devices / _cameras / _video / _arm mixins, each subclassing the base.
api/__init__ assembles AjaxRestApi(_HubsMixin, _DevicesMixin, _CamerasMixin,
_VideoMixin, _ArmMixin) and re-exports the public surface (AjaxRestApi + the
AjaxRest*Error exceptions + constants), so existing imports are unchanged.

Pure refactor: no behaviour, endpoint or unique_id change (verified by the full
suite + a live HA dev fetch). Tests patching asyncio.sleep / aiohttp.ClientSession
/ TCPConnector now target custom_components.ajax.api._base.*
