+++ custom_components/ajax/api/_base.py
@@ -1,4 +1,9 @@
-"""Ajax REST API client based on official PDF documentation."""
+"""Ajax REST API client — shared transport core (session, auth, retry, request).
+
+Houses ``AjaxRestClientBase`` plus the module-level constants and typed
+exceptions. Owns ``asyncio`` / ``aiohttp`` (backoff sleeps and session
+creation live here, which the auth tests patch).
+"""
 
 from __future__ import annotations
 
@@ -10,11 +15,11 @@ import logging
 import time
 from email.utils import parsedate_to_datetime
 from typing import Any
-from urllib.parse import urlencode, urlsplit
+from urllib.parse import urlsplit
 
 import aiohttp
 
-from .const import (
+from ..const import (
     AJAX_REST_API_BASE_URL,
     AJAX_REST_API_TIMEOUT,
     AUTH_MODE_DIRECT,
@@ -22,7 +27,7 @@ from .const import (
     INTEGRATION_VERSION,
 )
 
-_LOGGER = logging.getLogger(__name__)
+_LOGGER = logging.getLogger("custom_components.ajax.api")
 
 # Retry configuration
 MAX_RETRIES = 3
@@ -78,19 +83,8 @@ class AjaxRest2FARequiredError(AjaxRestApiError):
         self.request_id = request_id
 
 
-class AjaxRestApi:
-    """Ajax REST API client.
-
-    Authentication as User (from PDF page 4-5):
-    - First login via Credentials: E-mail + SHA256(password)
-    - Generate a temporary token
-    - Use this token for subsequent API calls
-
-    Supports three authentication modes:
-    - Direct: Connect directly to Ajax API with API key
-    - Proxy Secure: All requests go through proxy (proxy adds API key)
-    - Proxy Hybrid: Login via proxy to get API key, then direct API calls
-    """
+class AjaxRestClientBase:
+    """Transport, authentication and request-retry core shared by the API mixins."""
 
     def __init__(
         self,
@@ -926,746 +920,6 @@ class AjaxRestApi:
             _LOGGER.error("API request to %s timed out after %d retries", endpoint, MAX_RETRIES)
             raise AjaxRestConnectionError("Request timeout") from err
 
-    # Hub methods
-    async def async_get_hubs(self) -> list[dict[str, Any]]:
-        """Get all hubs.
-
-        Returns:
-            List of hub dictionaries
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-        return await self._request("GET", f"user/{self.user_id}/hubs")  # type: ignore[no-any-return]
-
-    async def async_get_hub(self, hub_id: str) -> dict[str, Any]:
-        """Get hub details.
-
-        Args:
-            hub_id: Hub ID
-
-        Returns:
-            Hub details dictionary
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-        return await self._request("GET", f"user/{self.user_id}/hubs/{hub_id}")  # type: ignore[no-any-return]
-
-    async def async_get_space_by_hub(self, hub_id: str) -> dict[str, Any] | None:
-        """Get space details by hub ID.
-
-        Args:
-            hub_id: Hub ID to find the associated space
-
-        Returns:
-            Space binding dictionary with id and name, or None if not found
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-        # Use query parameter to find space by hubId
-        spaces = await self._request("GET", f"user/{self.user_id}/spaces?hubId={hub_id}")
-        # Returns array of SpaceBinding, get first one
-        if spaces and isinstance(spaces, list) and len(spaces) > 0:
-            return spaces[0]  # type: ignore[no-any-return]
-        return None
-
-    async def async_get_rooms(self, hub_id: str) -> list[dict[str, Any]]:
-        """Get rooms for a hub.
-
-        Args:
-            hub_id: Hub ID
-
-        Returns:
-            List of room dictionaries
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-        return await self._request("GET", f"user/{self.user_id}/hubs/{hub_id}/rooms")  # type: ignore[no-any-return]
-
-    async def async_get_users(self, hub_id: str) -> list[dict[str, Any]]:
-        """Get users for a hub.
-
-        Args:
-            hub_id: Hub ID
-
-        Returns:
-            List of user dictionaries
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-        return await self._request("GET", f"user/{self.user_id}/hubs/{hub_id}/users")  # type: ignore[no-any-return]
-
-    async def async_get_hub_mode(self, hub_id: str) -> dict[str, Any]:
-        """Get hub alarm mode.
-
-        Args:
-            hub_id: Hub ID
-
-        Returns:
-            Hub mode dictionary
-        """
-        return await self._request("GET", f"hubs/{hub_id}/mode")  # type: ignore[no-any-return]
-
-    async def async_set_hub_mode(self, hub_id: str, mode: str) -> dict[str, Any]:
-        """Set hub alarm mode.
-
-        Args:
-            hub_id: Hub ID
-            mode: Alarm mode (full, partial_1, night, disarmed)
-
-        Returns:
-            Updated hub mode
-        """
-        return await self._request("POST", f"hubs/{hub_id}/mode", {"mode": mode})  # type: ignore[no-any-return]
-
-    # Device methods
-    async def async_get_devices(self, hub_id: str, enrich: bool = True) -> list[dict[str, Any]]:
-        """Get all devices for a specific hub.
-
-        Args:
-            hub_id: Hub ID
-            enrich: If True, returns full device details (default True)
-
-        Returns:
-            List of device dictionaries (with full details if enrich=True)
-        """
-        # Serve from the short-lived cache unless a cache bypass is pending
-        # (bypass must reach the real endpoint to get fresh state).
-        cache_key = (hub_id, enrich)
-        if not self._cache_bypass_active():
-            cached = self._devices_cache.get(cache_key)
-            if cached and (time.time() - cached[0]) < self._devices_cache_ttl:
-                return cached[1]
-
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-        endpoint = f"user/{self.user_id}/hubs/{hub_id}/devices"
-        if enrich:
-            endpoint += "?enrich=true"
-        data = await self._request("GET", endpoint)
-        self._devices_cache[cache_key] = (time.time(), data)
-        return data  # type: ignore[no-any-return]
-
-    async def async_get_device(self, hub_id: str, device_id: str) -> dict[str, Any]:
-        """Get device details.
-
-        Args:
-            hub_id: Hub ID
-            device_id: Device ID
-
-        Returns:
-            Device details dictionary
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-        return await self._request("GET", f"user/{self.user_id}/hubs/{hub_id}/devices/{device_id}")  # type: ignore[no-any-return]
-
-    async def async_get_device_state(self, device_id: str) -> dict[str, Any]:
-        """Get device state.
-
-        Args:
-            device_id: Device ID
-
-        Returns:
-            Device state dictionary
-        """
-        return await self._request("GET", f"devices/{device_id}/state")  # type: ignore[no-any-return]
-
-    async def async_send_device_command(self, hub_id: str, device_id: str, command: str, device_type: str) -> None:
-        """Send command to device (Socket/Relay/WallSwitch).
-
-        Uses the /command endpoint which is simpler and more reliable than PUT.
-        Supported commands: SWITCH_ON, SWITCH_OFF
-
-        Args:
-            hub_id: Hub ID
-            device_id: Device ID
-            command: Command string (e.g., "SWITCH_ON", "SWITCH_OFF")
-            device_type: Device type string (e.g., "WALL_SWITCH", "SOCKET", "RELAY")
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-
-        endpoint = f"user/{self.user_id}/hubs/{hub_id}/devices/{device_id}/command"
-        payload = {"command": command, "deviceType": device_type}
-        _LOGGER.info(
-            "Sending device command: POST %s with %s",
-            endpoint,
-            payload,
-        )
-        await self._request_no_response("POST", endpoint, payload)
-
-    async def async_set_switch_state(self, hub_id: str, device_id: str, state: bool, device_type: str) -> None:
-        """Set switch/relay/socket state.
-
-        Uses the /command endpoint for reliable switch control.
-
-        Args:
-            hub_id: Hub ID
-            device_id: Device ID
-            state: True for on, False for off
-            device_type: Device type string (e.g., "WallSwitch", "Socket", "Relay")
-        """
-        command = "SWITCH_ON" if state else "SWITCH_OFF"
-        await self.async_send_device_command(hub_id, device_id, command, device_type)
-
-    async def async_set_channel_state(
-        self,
-        hub_id: str,
-        device_id: str,
-        channel: int,
-        state: bool,
-        device_type: str,
-    ) -> None:
-        """Set multi-gang switch channel state.
-
-        Uses SWITCH_ON/SWITCH_OFF command with additionalParam.channels for LightSwitch.
-        The Ajax API expects a single channel per command.
-
-        Args:
-            hub_id: Hub ID
-            device_id: Device ID
-            channel: Channel index (0-based: 0 for channel 1, 1 for channel 2)
-            state: True for on, False for off
-            device_type: Device type string (e.g., "LightSwitchTwoGang")
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-
-        endpoint = f"user/{self.user_id}/hubs/{hub_id}/devices/{device_id}/command"
-        command = "SWITCH_ON" if state else "SWITCH_OFF"
-        # Convert 0-based index to API format: 0 -> "CHANNEL_1", 1 -> "CHANNEL_2"
-        channel_str = f"CHANNEL_{channel + 1}"
-        payload = {
-            "command": command,
-            "deviceType": device_type,
-            "additionalParam": {
-                "additionalParamType": "CHANNELS",
-                "channels": [channel_str],
-            },
-        }
-        _LOGGER.info(
-            "Sending channel command: POST %s with %s",
-            endpoint,
-            payload,
-        )
-        await self._request_no_response("POST", endpoint, payload)
-
-    async def async_set_waterstop_state(self, hub_id: str, device_id: str, open_valve: bool) -> None:
-        """Set WaterStop valve state.
-
-        Uses the /command endpoint for valve control.
-        WaterStop uses SWITCH_ON (open) / SWITCH_OFF (close) commands.
-
-        Args:
-            hub_id: Hub ID
-            device_id: Device ID
-            open_valve: True to open valve, False to close
-        """
-        # WaterStop uses same commands as switches: SWITCH_ON = open, SWITCH_OFF = close
-        command = "SWITCH_ON" if open_valve else "SWITCH_OFF"
-        await self.async_send_device_command(hub_id, device_id, command, "WaterStop")
-
-    # Socket methods
-    async def async_get_socket_power(self, device_id: str) -> dict[str, Any]:
-        """Get socket power consumption.
-
-        Args:
-            device_id: Socket device ID
-
-        Returns:
-            Power consumption data
-        """
-        return await self._request("GET", f"devices/{device_id}/power")  # type: ignore[no-any-return]
-
-    # Camera methods
-    async def async_get_cameras(self, hub_id: str) -> list[dict[str, Any]]:
-        """Get all cameras for a specific hub.
-
-        Args:
-            hub_id: Hub ID
-
-        Returns:
-            List of camera dictionaries
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-        return await self._request("GET", f"user/{self.user_id}/hubs/{hub_id}/cameras")  # type: ignore[no-any-return]
-
-    async def async_get_camera(self, hub_id: str, camera_id: str) -> dict[str, Any]:
-        """Get camera details.
-
-        Args:
-            hub_id: Hub ID
-            camera_id: Camera ID
-
-        Returns:
-            Camera details dictionary
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-        return await self._request("GET", f"user/{self.user_id}/hubs/{hub_id}/cameras/{camera_id}")  # type: ignore[no-any-return]
-
-    async def async_get_camera_snapshot(self, hub_id: str, camera_id: str) -> bytes:
-        """Get camera snapshot.
-
-        Args:
-            hub_id: Hub ID
-            camera_id: Camera ID
-
-        Returns:
-            Snapshot image data as bytes
-        """
-        if not self.session_token:
-            raise AjaxRestApiError("Not logged in. Call async_login() first.")
-
-        await self._proactive_token_refresh()
-
-        url = f"{self._get_base_url()}/user/{self.user_id}/hubs/{hub_id}/cameras/{camera_id}/snapshot"
-        session = await self._get_session()
-
-        headers: dict[str, str] = {k: v for k, v in self._base_headers.items() if v is not None}
-        headers["X-Session-Token"] = self.session_token
-
-        # Capture token version BEFORE the request to detect concurrent refresh
-        token_version_before = self._token_version
-
-        try:
-            async with session.get(
-                url,
-                headers=headers,
-                timeout=aiohttp.ClientTimeout(total=AJAX_REST_API_TIMEOUT),
-            ) as response:
-                if response.status == 401:
-                    # Try to recover auth and retry once
-                    async with self._auth_lock:
-                        if self._token_version != token_version_before:
-                            _LOGGER.debug("Token already refreshed by another coroutine")
-                        else:
-                            await self._recover_auth()
-                    headers["X-Session-Token"] = self.session_token
-                    async with session.get(
-                        url,
-                        headers=headers,
-                        timeout=aiohttp.ClientTimeout(total=AJAX_REST_API_TIMEOUT),
-                    ) as retry_response:
-                        retry_response.raise_for_status()
-                        return await retry_response.read()
-
-                response.raise_for_status()
-                return await response.read()
-        except aiohttp.ClientError as err:
-            raise AjaxRestConnectionError(f"Camera snapshot failed: {err}") from err
-        except TimeoutError as err:
-            raise AjaxRestConnectionError("Camera snapshot timeout") from err
-
-    async def async_get_camera_stream_url(self, hub_id: str, camera_id: str) -> str:
-        """Get camera stream URL.
-
-        Args:
-            hub_id: Hub ID
-            camera_id: Camera ID
-
-        Returns:
-            Stream URL string
-        """
-        data = await self._request("GET", f"user/{self.user_id}/hubs/{hub_id}/cameras/{camera_id}/stream")
-        return data.get("url", "")  # type: ignore[no-any-return]
-
-    # Video Edge methods (Bullet/Turret/MiniDome cameras)
-    async def async_get_space(self, space_id: str) -> dict[str, Any]:
-        """Get full space details including devices list.
-
-        Args:
-            space_id: Space ID (not hub_id)
-
-        Returns:
-            Space dictionary with devices array
-        """
-        # Skip the in-memory cache while a bypass window is open so the fresh
-        # space payload (which feeds video-edges and smart-locks) is fetched.
-        if not self._cache_bypass_active():
-            cached = self._space_cache.get(space_id)
-            if cached and (time.time() - cached[0]) < self._space_cache_ttl:
-                return cached[1]
-        data = await self._request("GET", f"user/{self.user_id}/spaces/{space_id}")
-        self._space_cache[space_id] = (time.time(), data)
-        return data  # type: ignore[no-any-return]
-
-    async def async_get_video_edges(self, space_id: str) -> list[dict[str, Any]]:
-        """Get all video edge devices for a space.
-
-        Args:
-            space_id: Space ID (real_space_id from spaceBinding)
-
-        Returns:
-            List of video edge dictionaries
-        """
-        # First get the space to find VIDEO_EDGE devices
-        space_data = await self.async_get_space(space_id)
-        devices = space_data.get("devices", [])
-        _LOGGER.debug(
-            "Space %s has %d devices: %s",
-            space_id,
-            len(devices),
-            [(d.get("id", "?")[:8], d.get("type")) for d in devices],
-        )
-        video_edges = []
-        for device in devices:
-            if device.get("type") == "VIDEO_EDGE":
-                video_edge_id = device.get("id")
-                if video_edge_id:
-                    try:
-                        video_edge = await self.async_get_video_edge(space_id, video_edge_id)
-                        video_edges.append(video_edge)
-                    except Exception as err:
-                        _LOGGER.warning("Failed to get video edge %s: %s", video_edge_id, err)
-        return video_edges
-
-    async def async_get_video_edge(self, space_id: str, video_edge_id: str) -> dict[str, Any]:
-        """Get video edge device details.
-
-        Args:
-            space_id: Space ID
-            video_edge_id: Video edge device ID
-
-        Returns:
-            Video edge details dictionary
-        """
-        return await self._request(  # type: ignore[no-any-return]
-            "GET",
-            f"user/{self.user_id}/spaces/{space_id}/devices/video-edges/{video_edge_id}",
-        )
-
-    async def async_get_video_edge_onvif(self, space_id: str, video_edge_id: str) -> dict[str, Any]:
-        """Get video edge ONVIF settings.
-
-        Args:
-            space_id: Space ID
-            video_edge_id: Video edge device ID
-
-        Returns:
-            ONVIF settings dictionary with userAuthEnabled, users, httpPort
-        """
-        return await self._request(  # type: ignore[no-any-return]
-            "GET",
-            f"user/{self.user_id}/spaces/{space_id}/devices/video-edges/{video_edge_id}/onvif",
-        )
-
-    async def async_get_video_edge_rtsp(self, space_id: str, video_edge_id: str) -> dict[str, Any]:
-        """Get video edge RTSP settings.
-
-        Args:
-            space_id: Space ID
-            video_edge_id: Video edge device ID
-
-        Returns:
-            RTSP settings dictionary with httpPort
-        """
-        return await self._request(  # type: ignore[no-any-return]
-            "GET",
-            f"user/{self.user_id}/spaces/{space_id}/devices/video-edges/{video_edge_id}/rtsp",
-        )
-
-    # Smart lock methods
-    async def async_get_smart_locks(self, space_id: str) -> list[dict[str, Any]]:
-        """Get all smart lock devices for a space.
-
-        Smart locks are discovered from the space's devices array
-        with type "SMART_LOCK", then fetched individually.
-
-        Args:
-            space_id: Space ID (real_space_id from spaceBinding)
-
-        Returns:
-            List of smart lock dictionaries
-        """
-        space_data = await self.async_get_space(space_id)
-        devices = space_data.get("devices", [])
-        _LOGGER.debug(
-            "Space %s devices for smart lock discovery: %s",
-            space_id,
-            [(d.get("id", "?")[:8], d.get("type")) for d in devices],
-        )
-        smart_locks = []
-        for device in devices:
-            if device.get("type") == "SMART_LOCK":
-                smart_lock_id = device.get("id")
-                if smart_lock_id:
-                    try:
-                        smart_lock = await self.async_get_smart_lock(space_id, smart_lock_id)
-                        # Merge name from space device listing if detail endpoint lacks it
-                        if not smart_lock.get("name") and device.get("name"):
-                            smart_lock["name"] = device["name"]
-                        smart_locks.append(smart_lock)
-                    except Exception as err:
-                        _LOGGER.warning("Failed to get smart lock %s: %s", smart_lock_id, err)
-                        # Fallback: use the space device entry directly
-                        smart_locks.append(device)
-        return smart_locks
-
-    async def async_get_smart_lock(self, space_id: str, smart_lock_id: str) -> dict[str, Any]:
-        """Get smart lock device details.
-
-        Args:
-            space_id: Space ID
-            smart_lock_id: Smart lock device ID
-
-        Returns:
-            Smart lock details dictionary
-        """
-        return await self._request(  # type: ignore[no-any-return]
-            "GET",
-            f"user/{self.user_id}/spaces/{space_id}/devices/smart-locks/{smart_lock_id}",
-        )
-
-    # Light/Button methods
-    async def async_set_dimmer_brightness(
-        self,
-        hub_id: str,
-        device_id: str,
-        brightness: int,
-    ) -> None:
-        """Set dimmer brightness level.
-
-        Args:
-            hub_id: Hub ID
-            device_id: Dimmer device ID
-            brightness: Brightness level (0-100). 0 = off.
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-
-        # Use command endpoint with BRIGHTNESS command
-        await self._request_no_response(
-            "POST",
-            f"user/{self.user_id}/hubs/{hub_id}/devices/{device_id}/command",
-            {
-                "command": "BRIGHTNESS",
-                "deviceType": "LightSwitchDimmer",
-                "additionalParam": {
-                    "additionalParamType": "BRIGHTNESS_STATUS",
-                    "brightnessInPercentage": brightness,
-                    "channels": ["CHANNEL_1"],
-                    "brightnessType": "BRIGHTNESS_TYPE_ABSOLUTE",
-                },
-            },
-        )
-
-    # Automation methods
-    async def async_get_automations(self, hub_id: str) -> list[dict[str, Any]]:
-        """Get hub automations/scenarios.
-
-        Args:
-            hub_id: Hub ID
-
-        Returns:
-            List of automation dictionaries
-        """
-        return await self._request("GET", f"hubs/{hub_id}/automations")  # type: ignore[no-any-return]
-
-    async def async_trigger_automation(
-        self,
-        hub_id: str,
-        automation_id: str,
-    ) -> dict[str, Any]:
-        """Trigger automation/scenario.
-
-        Args:
-            hub_id: Hub ID
-            automation_id: Automation ID
-
-        Returns:
-            Automation trigger response
-        """
-        return await self._request("POST", f"hubs/{hub_id}/automations/{automation_id}/trigger")  # type: ignore[no-any-return]
-
-    # Events methods
-    async def async_get_events(self, hub_id: str, limit: int = 100) -> list[dict[str, Any]]:
-        """Get hub events history.
-
-        Args:
-            hub_id: Hub ID
-            limit: Maximum number of events to return
-
-        Returns:
-            List of event dictionaries
-        """
-        return await self._request("GET", f"hubs/{hub_id}/events?limit={limit}")  # type: ignore[no-any-return]
-
-    # NVR methods
-    async def async_get_nvr_recordings(
-        self,
-        nvr_id: str,
-        camera_id: str,
-        start: str,
-        end: str,
-    ) -> list[dict[str, Any]]:
-        """Get NVR recordings for a camera.
-
-        Args:
-            nvr_id: NVR device ID
-            camera_id: Camera ID
-            start: Start time (ISO format)
-            end: End time (ISO format)
-
-        Returns:
-            List of recording dictionaries
-        """
-        query = urlencode({"cameraId": camera_id, "start": start, "end": end})
-        return await self._request("GET", f"devices/{nvr_id}/recordings?{query}")  # type: ignore[no-any-return]
-
-    # Fields that should not be sent in device update requests (read-only or internal)
-    # Note: deviceType MUST be included - API needs it for deserialization
-    DEVICE_UPDATE_EXCLUDE_FIELDS = {
-        # Identifiers (read-only)
-        "id",
-        "hubId",
-        "spaceId",
-        "serial",
-        # Type info (read-only) - but NOT deviceType (required by API)
-        "type",
-        "deviceCategory",
-        # Version info (read-only)
-        "firmwareVersion",
-        "hardwareVersion",
-        # State/status fields (computed)
-        "online",
-        "state",
-        "status",
-        "batteryLevel",
-        "batteryChargeLevelPercentage",
-        "signalLevel",
-        # Sensor state fields (read-only)
-        "temperature",
-        "tampered",
-        "reedClosed",
-        "extraContactClosed",
-        "externalContactState",
-        "externalContactTriggered",
-        "estimatedArmingState",
-        "issuesCount",
-        "confirmsAlarm",
-        "verifiesAlarm",
-        # Timestamps (read-only)
-        "createdAt",
-        "updatedAt",
-        "lastSeen",
-        "lastSeenAt",
-        # Computed/internal fields that cause API validation errors
-        "wiringSchemeSpecificDetails",
-        "malfunctions",
-        "alerts",
-        "problems",
-        # Device capability/support flags (read-only)
-        "capabilities",
-        "selfMonitoringConfigs",
-        "indicatorLightModeSupported",
-        "rollerShutterSupported",
-        # Hardware config fields (read-only)
-        "color",
-        # Note: assignedExtender is required by API, do not exclude
-        "cmsDeviceIndex",
-        "bypassState",
-        # Note: deviceTransmissionPowerMode/Value must NOT be excluded -
-        # API requires at least one to be present in PUT requests
-    }
-
-    # Device settings methods
-    async def async_update_device(
-        self,
-        hub_id: str,
-        device_id: str,
-        settings: dict[str, Any],
-    ) -> None:
-        """Update device settings.
-
-        Args:
-            hub_id: Hub ID
-            device_id: Device ID
-            settings: Dictionary of settings to update (e.g., {"alwaysActive": true})
-
-        Raises:
-            AjaxRestApiError: If the update fails
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-
-        # First get current device data
-        current_device = await self.async_get_device(hub_id, device_id)
-
-        # Merge settings with current device data
-        updated_device = {**current_device, **settings}
-
-        # Remove read-only and problematic fields
-        for field in self.DEVICE_UPDATE_EXCLUDE_FIELDS:
-            updated_device.pop(field, None)
-
-        _LOGGER.debug(
-            "Updating device %s with fields: %s",
-            device_id,
-            list(updated_device.keys()),
-        )
-
-        await self._request_no_response(
-            "PUT",
-            f"user/{self.user_id}/hubs/{hub_id}/devices/{device_id}",
-            updated_device,
-        )
-
-    async def async_update_device_nested(
-        self,
-        hub_id: str,
-        device_id: str,
-        settings: dict[str, Any],
-    ) -> None:
-        """Update device settings with deep merge for nested structures.
-
-        This method properly handles nested settings like wiredDeviceSettings
-        by merging them with existing values instead of replacing.
-
-        Args:
-            hub_id: Hub ID
-            device_id: Device ID
-            settings: Dictionary of settings to update, can include nested dicts
-
-        Raises:
-            AjaxRestApiError: If the update fails
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-
-        # First get current device data
-        current_device = await self.async_get_device(hub_id, device_id)
-
-        # Deep merge settings with current device data
-        updated_device = self._deep_merge(current_device, settings)
-
-        # Remove read-only and problematic fields
-        for field in self.DEVICE_UPDATE_EXCLUDE_FIELDS:
-            updated_device.pop(field, None)
-
-        _LOGGER.debug(
-            "Updating device %s (nested) with fields: %s",
-            device_id,
-            list(updated_device.keys()),
-        )
-
-        await self._request_no_response(
-            "PUT",
-            f"user/{self.user_id}/hubs/{hub_id}/devices/{device_id}",
-            updated_device,
-        )
-
-    def _deep_merge(self, base: dict[str, Any], updates: dict[str, Any]) -> dict[str, Any]:
-        """Deep merge two dictionaries, preserving nested structures."""
-        result = base.copy()
-        for key, value in updates.items():
-            if key in result and isinstance(result[key], dict) and isinstance(value, dict):
-                result[key] = self._deep_merge(result[key], value)
-            else:
-                result[key] = value
-        return result
-
     async def _request_no_response(
         self,
         method: str,
@@ -1815,120 +1069,3 @@ class AjaxRestApi:
                 return await self._request_no_response(method, endpoint, data, _retry_on_auth_error, _retry_count + 1)
             _LOGGER.error("API request to %s timed out after %d retries", endpoint, MAX_RETRIES)
             raise AjaxRestConnectionError("Request timeout") from err
-
-    # Arming commands
-    async def async_arm(self, hub_id: str, ignore_problems: bool = True) -> None:
-        """Arm the hub.
-
-        Args:
-            hub_id: Hub ID
-            ignore_problems: Whether to ignore sensor problems when arming
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-
-        await self._request_no_response(
-            "PUT",
-            f"user/{self.user_id}/hubs/{hub_id}/commands/arming",
-            {"command": "ARM", "ignoreProblems": ignore_problems},
-        )
-
-    async def async_disarm(self, hub_id: str, ignore_problems: bool = True) -> None:
-        """Disarm the hub.
-
-        Args:
-            hub_id: Hub ID
-            ignore_problems: Whether to ignore sensor problems
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-
-        await self._request_no_response(
-            "PUT",
-            f"user/{self.user_id}/hubs/{hub_id}/commands/arming",
-            {"command": "DISARM", "ignoreProblems": ignore_problems},
-        )
-
-    async def async_night_mode(self, hub_id: str, enabled: bool = True) -> None:
-        """Set night mode on/off.
-
-        Args:
-            hub_id: Hub ID
-            enabled: True for night mode on, False for off
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-
-        command = "NIGHT_MODE_ON" if enabled else "NIGHT_MODE_OFF"
-        await self._request_no_response(
-            "PUT",
-            f"user/{self.user_id}/hubs/{hub_id}/commands/arming",
-            {"command": command, "ignoreProblems": True},
-        )
-
-    async def async_press_panic_button(self, hub_id: str) -> None:
-        """Trigger panic alarm for a hub.
-
-        Args:
-            hub_id: Hub ID
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-
-        await self._request_no_response(
-            "PUT",
-            f"user/{self.user_id}/hubs/{hub_id}/commands/arming",
-            {"command": "PANIC"},
-        )
-
-    # Group commands
-    async def async_get_groups(self, hub_id: str) -> list[dict[str, Any]]:
-        """Get all groups for a hub.
-
-        Args:
-            hub_id: Hub ID
-
-        Returns:
-            List of group data dictionaries
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-
-        return await self._request(  # type: ignore[no-any-return]
-            "GET",
-            f"user/{self.user_id}/hubs/{hub_id}/groups",
-        )
-
-    async def async_arm_group(self, hub_id: str, group_id: str, ignore_problems: bool = True) -> None:
-        """Arm a specific group.
-
-        Args:
-            hub_id: Hub ID
-            group_id: Group ID to arm
-            ignore_problems: Whether to ignore sensor problems when arming
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-
-        await self._request_no_response(
-            "PUT",
-            f"user/{self.user_id}/hubs/{hub_id}/groups/{group_id}/commands/arming",
-            {"command": "ARM", "ignoreProblems": ignore_problems},
-        )
-
-    async def async_disarm_group(self, hub_id: str, group_id: str, ignore_problems: bool = True) -> None:
-        """Disarm a specific group.
-
-        Args:
-            hub_id: Hub ID
-            group_id: Group ID to disarm
-            ignore_problems: Whether to ignore sensor problems
-        """
-        if not self.user_id:
-            raise AjaxRestApiError("No user_id available. Call async_login() first.")
-
-        await self._request_no_response(
-            "PUT",
-            f"user/{self.user_id}/hubs/{hub_id}/groups/{group_id}/commands/arming",
-            {"command": "DISARM", "ignoreProblems": ignore_problems},
-        )
